+++ changelog.md
@@ -21,6 +21,7 @@ If you are reading this on the website, then these changes are live here.
 - Error and diagnostic messages a no-longer all one color
 - The pad editor now inserts a trailing newline on format
 - Increase thresholds for arrays becoming images or audio
+- Split up system functions into more categories on the main docs page
 
 ## 0.0.22 - 2023-10-21
 ### Language